🐛 fix: fileinfoではなく、filestatに戻した
